refactor(ecies): move ECIES Codec to separate file
