+++ crates/net/ecies/src/proto.rs
@@ -16,150 +16,10 @@ use tokio::{
     net::TcpStream,
 };
 use tokio_stream::{Stream, StreamExt};
-use tokio_util::codec::{Decoder, Encoder, Framed};
+use tokio_util::codec::{Decoder, Framed};
 use tracing::{debug, instrument, trace};
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-/// Current ECIES state of a connection
-pub enum ECIESState {
-    /// The first stage of the ECIES handshake, where each side of the connection sends an auth
-    /// message containing the ephemeral public key, signature of the public key, nonce, and other
-    /// metadata.
-    Auth,
-
-    /// The second stage of the ECIES handshake, where each side of the connection sends an ack
-    /// message containing the nonce and other metadata.
-    Ack,
-    Header,
-    Body,
-}
-
-/// Tokio codec for ECIES
-#[derive(Debug)]
-pub struct ECIESCodec {
-    ecies: ECIES,
-    state: ECIESState,
-}
-
-impl ECIESCodec {
-    /// Create a new server codec using the given secret key
-    pub fn new_server(secret_key: SecretKey) -> Result<Self, ECIESError> {
-        Ok(Self { ecies: ECIES::new_server(secret_key)?, state: ECIESState::Auth })
-    }
-
-    /// Create a new client codec using the given secret key and the server's public id
-    pub fn new_client(secret_key: SecretKey, remote_id: PeerId) -> Result<Self, ECIESError> {
-        Ok(Self { ecies: ECIES::new_client(secret_key, remote_id)?, state: ECIESState::Auth })
-    }
-}
-
-impl Decoder for ECIESCodec {
-    type Item = IngressECIESValue;
-    type Error = ECIESError;
-
-    #[instrument(level = "trace", skip_all, fields(peer=&*format!("{:?}", self.ecies.remote_id.map(|s| s.to_string())), state=&*format!("{:?}", self.state)))]
-    fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
-        loop {
-            match self.state {
-                ECIESState::Auth => {
-                    trace!("parsing auth");
-                    if buf.len() < 2 {
-                        return Ok(None)
-                    }
-
-                    let payload_size = u16::from_be_bytes([buf[0], buf[1]]) as usize;
-                    let total_size = payload_size + 2;
-
-                    if buf.len() < total_size {
-                        trace!("current len {}, need {}", buf.len(), total_size);
-                        return Ok(None)
-                    }
-
-                    self.ecies.read_auth(&mut *buf.split_to(total_size))?;
-
-                    self.state = ECIESState::Header;
-                    return Ok(Some(IngressECIESValue::AuthReceive(self.ecies.remote_id())))
-                }
-                ECIESState::Ack => {
-                    trace!("parsing ack with len {}", buf.len());
-                    if buf.len() < 2 {
-                        return Ok(None)
-                    }
-
-                    let payload_size = u16::from_be_bytes([buf[0], buf[1]]) as usize;
-                    let total_size = payload_size + 2;
-
-                    if buf.len() < total_size {
-                        trace!("current len {}, need {}", buf.len(), total_size);
-                        return Ok(None)
-                    }
-
-                    self.ecies.read_ack(&mut *buf.split_to(total_size))?;
-
-                    self.state = ECIESState::Header;
-                    return Ok(Some(IngressECIESValue::Ack))
-                }
-                ECIESState::Header => {
-                    if buf.len() < ECIES::header_len() {
-                        trace!("current len {}, need {}", buf.len(), ECIES::header_len());
-                        return Ok(None)
-                    }
-
-                    self.ecies.read_header(&mut *buf.split_to(ECIES::header_len()))?;
-
-                    self.state = ECIESState::Body;
-                }
-                ECIESState::Body => {
-                    if buf.len() < self.ecies.body_len() {
-                        return Ok(None)
-                    }
-
-                    let mut data = buf.split_to(self.ecies.body_len());
-                    let ret = Bytes::copy_from_slice(self.ecies.read_body(&mut *data)?);
-
-                    self.state = ECIESState::Header;
-                    return Ok(Some(IngressECIESValue::Message(ret)))
-                }
-            }
-        }
-    }
-}
-
-impl Encoder<EgressECIESValue> for ECIESCodec {
-    type Error = io::Error;
-
-    #[instrument(level = "trace", skip(self, buf), fields(peer=&*format!("{:?}", self.ecies.remote_id.map(|s| s.to_string())), state=&*format!("{:?}", self.state)))]
-    fn encode(&mut self, item: EgressECIESValue, buf: &mut BytesMut) -> Result<(), Self::Error> {
-        match item {
-            EgressECIESValue::Auth => {
-                self.state = ECIESState::Ack;
-                self.ecies.write_auth(buf);
-                Ok(())
-            }
-            EgressECIESValue::Ack => {
-                self.state = ECIESState::Header;
-                self.ecies.write_ack(buf);
-                Ok(())
-            }
-            EgressECIESValue::Message(data) => {
-                if data.len() > MAX_BODY_SIZE {
-                    return Err(io::Error::new(
-                        io::ErrorKind::InvalidInput,
-                        format!(
-                            "body size ({}) exceeds limit ({} bytes)",
-                            data.len(),
-                            MAX_BODY_SIZE
-                        ),
-                    ))
-                }
-
-                self.ecies.write_header(buf, data.len());
-                self.ecies.write_body(buf, &data);
-                Ok(())
-            }
-        }
-    }
-}
+use crate::codec::ECIESCodec;
 
 /// `ECIES` stream over TCP exchanging raw bytes
 #[derive(Debug)]
@@ -168,8 +28,9 @@ pub struct ECIESStream<Io> {
     remote_id: PeerId,
 }
 
-// This trait is just for instrumenting the stream with a socket addr
+/// This trait is just for instrumenting the stream with a socket addr
 pub trait HasRemoteAddr {
+    /// Maybe returns a [`SocketAddr`]
     fn remote_addr(&self) -> Option<SocketAddr>;
 }
 
@@ -221,7 +82,7 @@ where
         debug!("receiving ecies auth");
         let remote_id = match &msg {
             Some(IngressECIESValue::AuthReceive(remote_id)) => *remote_id,
-            other => {
+            _ => {
                 return Err(ECIESError::InvalidHandshake {
                     expected: IngressECIESValue::AuthReceive(Default::default()),
                     msg,
